Implemented saving the results of the client-server operations to a file.
Added a size limit for console display of a matrix, as well as added
ability to save the operation stats to the file by modifying the method
parameters and how we store those stats.

Code cleanup and documented any missing documentation.

Added all of our names to the author tag of each file.